Restructure the test directory
* Restructure the test directory introducing integration and local
  directories
