+++ test/local/fixed-rates-calculator.test.ts
@@ -2,10 +2,10 @@ import {ethers, waffle} from 'hardhat'
 import chai from 'chai'
 import {solidity} from "ethereum-waffle";
 
-import FixedRatesCalculatorArtifact from '../artifacts/contracts/FixedRatesCalculator.sol/FixedRatesCalculator.json';
-import {FixedRatesCalculator} from '../typechain/FixedRatesCalculator';
+import FixedRatesCalculatorArtifact from '../../artifacts/contracts/FixedRatesCalculator.sol/FixedRatesCalculator.json';
+import {FixedRatesCalculator} from '../../typechain/FixedRatesCalculator';
 import {SignerWithAddress} from "@nomiclabs/hardhat-ethers/signers";
-import {fromWei, getFixedGasSigners, toWei} from "./_helpers";
+import {fromWei, getFixedGasSigners, toWei} from "../_helpers";
 
 chai.use(solidity);
