+++ test/local/pool-fixed-rates-borrowing.test.ts
@@ -2,14 +2,14 @@ import {ethers, waffle} from 'hardhat'
 import chai, {expect} from 'chai'
 import {solidity} from "ethereum-waffle";
 
-import FixedRatesCalculatorArtifact from '../artifacts/contracts/FixedRatesCalculator.sol/FixedRatesCalculator.json';
-import PoolArtifact from '../artifacts/contracts/Pool.sol/Pool.json';
+import FixedRatesCalculatorArtifact from '../../artifacts/contracts/FixedRatesCalculator.sol/FixedRatesCalculator.json';
+import PoolArtifact from '../../artifacts/contracts/Pool.sol/Pool.json';
 import OpenBorrowersRegistryArtifact
-  from '../artifacts/contracts/OpenBorrowersRegistry.sol/OpenBorrowersRegistry.json';
+  from '../../artifacts/contracts/OpenBorrowersRegistry.sol/OpenBorrowersRegistry.json';
 import {SignerWithAddress} from "@nomiclabs/hardhat-ethers/signers";
-import {fromWei, getFixedGasSigners, time, toWei} from "./_helpers";
-import {FixedRatesCalculator, OpenBorrowersRegistry, Pool} from "../typechain";
-import {CompoundingIndex__factory} from "../typechain";
+import {fromWei, getFixedGasSigners, time, toWei} from "../_helpers";
+import {FixedRatesCalculator, OpenBorrowersRegistry, Pool} from "../../typechain";
+import {CompoundingIndex__factory} from "../../typechain";
 
 chai.use(solidity);
