+++ test/local/pool-flexible-rates-deposits.test.ts
@@ -2,12 +2,12 @@ import {ethers, waffle} from 'hardhat'
 import chai, {expect} from 'chai'
 import {solidity} from "ethereum-waffle";
 
-import FixedRatesCalculatorArtifact from '../artifacts/contracts/FixedRatesCalculator.sol/FixedRatesCalculator.json';
-import PoolArtifact from '../artifacts/contracts/Pool.sol/Pool.json';
+import FixedRatesCalculatorArtifact from '../../artifacts/contracts/FixedRatesCalculator.sol/FixedRatesCalculator.json';
+import PoolArtifact from '../../artifacts/contracts/Pool.sol/Pool.json';
 import {SignerWithAddress} from "@nomiclabs/hardhat-ethers/signers";
-import {fromWei, getFixedGasSigners, time, toWei} from "./_helpers";
-import {FixedRatesCalculator, Pool} from "../typechain";
-import {CompoundingIndex__factory, OpenBorrowersRegistry__factory} from "../typechain";
+import {fromWei, getFixedGasSigners, time, toWei} from "../_helpers";
+import {FixedRatesCalculator, Pool} from "../../typechain";
+import {CompoundingIndex__factory, OpenBorrowersRegistry__factory} from "../../typechain";
 
 chai.use(solidity);
