+++ test/local/pool-utilisation-rates.test.ts
@@ -3,14 +3,14 @@ import chai, {expect} from 'chai'
 import {solidity} from "ethereum-waffle";
 
 import UtilisationRatesCalculatorArtifact
-  from '../artifacts/contracts/UtilisationRatesCalculator.sol/UtilisationRatesCalculator.json';
-import PoolArtifact from '../artifacts/contracts/Pool.sol/Pool.json';
+  from '../../artifacts/contracts/UtilisationRatesCalculator.sol/UtilisationRatesCalculator.json';
+import PoolArtifact from '../../artifacts/contracts/Pool.sol/Pool.json';
 import OpenBorrowersRegistryArtifact
-  from '../artifacts/contracts/OpenBorrowersRegistry.sol/OpenBorrowersRegistry.json';
+  from '../../artifacts/contracts/OpenBorrowersRegistry.sol/OpenBorrowersRegistry.json';
 import {SignerWithAddress} from "@nomiclabs/hardhat-ethers/signers";
-import {fromWei, getFixedGasSigners, time, toWei} from "./_helpers";
-import {OpenBorrowersRegistry, Pool, UtilisationRatesCalculator} from "../typechain";
-import {CompoundingIndex__factory, OpenBorrowersRegistry__factory} from "../typechain";
+import {fromWei, getFixedGasSigners, time, toWei} from "../_helpers";
+import {OpenBorrowersRegistry, Pool, UtilisationRatesCalculator} from "../../typechain";
+import {CompoundingIndex__factory, OpenBorrowersRegistry__factory} from "../../typechain";
 
 chai.use(solidity);
