+++ test/local/simple-price-provider.test.ts
@@ -2,10 +2,10 @@ import {ethers, waffle} from 'hardhat'
 import chai from 'chai'
 import {solidity} from "ethereum-waffle";
 
-import SimplePriceProviderArtifact from '../artifacts/contracts/SimplePriceProvider.sol/SimplePriceProvider.json';
+import SimplePriceProviderArtifact from '../../artifacts/contracts/SimplePriceProvider.sol/SimplePriceProvider.json';
 import {SignerWithAddress} from "@nomiclabs/hardhat-ethers/signers";
-import {fromWei, getFixedGasSigners, toBytes32, toWei} from "./_helpers";
-import {SimplePriceProvider} from "../typechain";
+import {fromWei, getFixedGasSigners, toBytes32, toWei} from "../_helpers";
+import {SimplePriceProvider} from "../../typechain";
 
 chai.use(solidity);
